assignment3/requirements.txt
tensorflow-cpu
matplotlib
seaborn
scikit-learn
pandas